backend/requirements.txt
djangorestframework==3.1.8
Pillow==8.2.0
crispy-forms-gds==0.2.2
yfinance==0.1.59
plotly